Support for named exports and arrays.
Any named es6 export on the loaded file is interpreted as class definition and appended to the definitions from the default export. This allows styles to be defined with exports that map 1:1 with css modules.

Any block, selector or property can now be defined as an array of values. This allows defining multiple properties of the same name within a class.